refactor(Testimonials): move Testimonial to a new file and better practices
